refactor: 훅으로 분리되어 있던 블럭 포커스 이동 로직 Mom 컴포넌트로 이동
